:truck: [#1579] Move FormVariableSources import
